+++ src/eval/chaining.rs
@@ -4,7 +4,7 @@
 use super::{Caches, GlobalRuntimeState, Target};
 use crate::ast::{ASTFlags, Expr, OpAssignment};
 use crate::types::dynamic::Union;
-use crate::{Dynamic, Engine, Module, Position, RhaiResult, RhaiResultOf, Scope, StaticVec, ERR};
+use crate::{Dynamic, Engine, FnArgsVec, Module, Position, RhaiResult, RhaiResultOf, Scope, ERR};
 use std::hash::Hash;
 #[cfg(feature = "no_std")]
 use std::prelude::v1::*;
@@ -33,92 +33,9 @@ impl From<&Expr> for ChainType {
     }
 }
 
-/// Value of a chaining argument.
-#[derive(Debug, Clone, Hash)]
-pub enum ChainArgument {
-    /// Dot-property access.
-    #[cfg(not(feature = "no_object"))]
-    Property(Position),
-    /// Arguments to a dot method call.
-    /// Wrapped values are the arguments plus the [position][Position] of the first argument.
-    ///
-    /// Since many dotted function calls have no arguments (e.g. `string.len()`), it is better to
-    /// reduce the size of [`ChainArgument`] by using a boxed slice.
-    #[cfg(not(feature = "no_object"))]
-    MethodCallArgs(Box<[Dynamic]>, Position),
-    /// Index value and [position][Position].
-    #[cfg(not(feature = "no_index"))]
-    IndexValue(Dynamic, Position),
-}
-
-impl ChainArgument {
-    /// Return the index value.
-    #[inline(always)]
-    #[cfg(not(feature = "no_index"))]
-    #[must_use]
-    pub fn into_index_value(self) -> Option<Dynamic> {
-        match self {
-            Self::IndexValue(value, ..) => Some(value),
-            #[cfg(not(feature = "no_object"))]
-            _ => None,
-        }
-    }
-    /// Return the list of method call arguments.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the [`ChainArgument`] is not [`MethodCallArgs`][ChainArgument::MethodCallArgs].
-    #[inline(always)]
-    #[cfg(not(feature = "no_object"))]
-    #[must_use]
-    pub fn into_fn_call_args(self) -> (crate::FnArgsVec<Dynamic>, Position) {
-        match self {
-            Self::MethodCallArgs(values, pos) if values.is_empty() => {
-                (crate::FnArgsVec::new_const(), pos)
-            }
-            Self::MethodCallArgs(mut values, pos) => {
-                (values.iter_mut().map(std::mem::take).collect(), pos)
-            }
-            x => unreachable!("ChainArgument::MethodCallArgs expected but gets {:?}", x),
-        }
-    }
-    /// Return the [position][Position].
-    #[inline(always)]
-    #[must_use]
-    #[allow(dead_code)]
-    pub const fn position(&self) -> Position {
-        match self {
-            #[cfg(not(feature = "no_object"))]
-            ChainArgument::Property(pos) => *pos,
-            #[cfg(not(feature = "no_object"))]
-            ChainArgument::MethodCallArgs(.., pos) => *pos,
-            #[cfg(not(feature = "no_index"))]
-            ChainArgument::IndexValue(.., pos) => *pos,
-        }
-    }
-    /// Create n [`MethodCallArgs`][ChainArgument::MethodCallArgs].
-    #[inline(always)]
-    #[cfg(not(feature = "no_object"))]
-    #[must_use]
-    pub fn from_fn_call_args(values: crate::FnArgsVec<Dynamic>, pos: Position) -> Self {
-        if values.is_empty() {
-            Self::MethodCallArgs(Box::default(), pos)
-        } else {
-            Self::MethodCallArgs(values.into_boxed_slice(), pos)
-        }
-    }
-    /// Create an [`IndexValue`][ChainArgument::IndexValue].
-    #[inline(always)]
-    #[cfg(not(feature = "no_index"))]
-    #[must_use]
-    pub const fn from_index_value(value: Dynamic, pos: Position) -> Self {
-        Self::IndexValue(value, pos)
-    }
-}
-
 impl Engine {
     /// Chain-evaluate a dot/index chain.
-    /// [`Position`] in [`EvalAltResult`] is [`NONE`][Position::NONE] and must be set afterwards.
+    /// [`Position`] in [`EvalAltResult`] may be [`NONE`][Position::NONE] and should be set afterwards.
     fn eval_dot_index_chain_helper(
         &self,
         global: &mut GlobalRuntimeState,
@@ -130,15 +47,12 @@ impl Engine {
         _parent: &Expr,
         rhs: &Expr,
         _parent_options: ASTFlags,
-        idx_values: &mut StaticVec<ChainArgument>,
+        idx_values: &mut FnArgsVec<Dynamic>,
         chain_type: ChainType,
         level: usize,
         new_val: Option<(Dynamic, OpAssignment)>,
     ) -> RhaiResultOf<(Dynamic, bool)> {
         let is_ref_mut = target.is_ref();
-
-        // Pop the last index value
-        let idx_val = idx_values.pop().unwrap();
 
         #[cfg(feature = "debugging")]
         let scope = &mut Scope::new();
@@ -147,7 +61,6 @@ impl Engine {
             #[cfg(not(feature = "no_index"))]
             ChainType::Indexing => {
                 let pos = rhs.start_position();
-                let idx_val = idx_val.into_index_value().expect("`ChainType::Index`");
 
                 match rhs {
                     // xxx[idx].expr... | xxx[idx][expr]...
@@ -157,6 +70,7 @@ impl Engine {
                         #[cfg(feature = "debugging")]
                         self.run_debugger(scope, global, lib, this_ptr, _parent, level)?;
 
+                        let idx_val = idx_values.pop().unwrap();
                         let mut idx_val_for_setter = idx_val.clone();
                         let idx_pos = x.lhs.start_position();
                         let rhs_chain = rhs.into();
@@ -201,6 +115,7 @@ impl Engine {
                         self.run_debugger(scope, global, lib, this_ptr, _parent, level)?;
 
                         let (new_val, op_info) = new_val.expect("`Some`");
+                        let idx_val = idx_values.pop().unwrap();
                         let mut idx_val2 = idx_val.clone();
 
                         let try_setter = match self.get_indexed_mut(
@@ -259,6 +174,8 @@ impl Engine {
                         #[cfg(feature = "debugging")]
                         self.run_debugger(scope, global, lib, this_ptr, _parent, level)?;
 
+                        let idx_val = idx_values.pop().unwrap();
+
                         self.get_indexed_mut(
                             global, caches, lib, target, idx_val, pos, false, true, level,
                         )
@@ -269,19 +186,32 @@ impl Engine {
 
             #[cfg(not(feature = "no_object"))]
             ChainType::Dotting => {
+                // Check for existence with the Elvis operator
+                if _parent_options.contains(ASTFlags::NEGATED) && target.is::<()>() {
+                    return Ok((Dynamic::UNIT, false));
+                }
+
                 match rhs {
                     // xxx.fn_name(arg_expr_list)
                     Expr::MethodCall(x, pos) if !x.is_qualified() && new_val.is_none() => {
-                        let crate::ast::FnCallExpr { name, hashes, .. } = x.as_ref();
-                        let call_args = &mut idx_val.into_fn_call_args();
-
                         #[cfg(feature = "debugging")]
                         let reset_debugger =
                             self.run_debugger_with_reset(scope, global, lib, this_ptr, rhs, level)?;
 
+                        let crate::ast::FnCallExpr {
+                            name, hashes, args, ..
+                        } = x.as_ref();
+
+                        let offset = idx_values.len() - args.len();
+                        let call_args = &mut idx_values[offset..];
+                        let pos1 = args.get(0).map_or(Position::NONE, Expr::position);
+
                         let result = self.make_method_call(
-                            global, caches, lib, name, *hashes, target, call_args, *pos, level,
+                            global, caches, lib, name, *hashes, target, call_args, pos1, *pos,
+                            level,
                         );
+
+                        idx_values.truncate(offset);
 
                         #[cfg(feature = "debugging")]
                         global.debugger.reset_status(reset_debugger);
@@ -440,18 +370,25 @@ impl Engine {
                             }
                             // {xxx:map}.fn_name(arg_expr_list)[expr] | {xxx:map}.fn_name(arg_expr_list).expr
                             Expr::MethodCall(ref x, pos) if !x.is_qualified() => {
-                                let crate::ast::FnCallExpr { name, hashes, .. } = x.as_ref();
-                                let call_args = &mut idx_val.into_fn_call_args();
-
                                 #[cfg(feature = "debugging")]
                                 let reset_debugger = self.run_debugger_with_reset(
                                     scope, global, lib, this_ptr, _node, level,
                                 )?;
 
+                                let crate::ast::FnCallExpr {
+                                    name, hashes, args, ..
+                                } = x.as_ref();
+
+                                let offset = idx_values.len() - args.len();
+                                let call_args = &mut idx_values[offset..];
+                                let pos1 = args.get(0).map_or(Position::NONE, Expr::position);
+
                                 let result = self.make_method_call(
-                                    global, caches, lib, name, *hashes, target, call_args, pos,
-                                    level,
+                                    global, caches, lib, name, *hashes, target, call_args, pos1,
+                                    pos, level,
                                 );
+
+                                idx_values.truncate(offset);
 
                                 #[cfg(feature = "debugging")]
                                 global.debugger.reset_status(reset_debugger);
@@ -558,18 +495,26 @@ impl Engine {
                             }
                             // xxx.fn_name(arg_expr_list)[expr] | xxx.fn_name(arg_expr_list).expr
                             Expr::MethodCall(ref f, pos) if !f.is_qualified() => {
-                                let crate::ast::FnCallExpr { name, hashes, .. } = f.as_ref();
-                                let rhs_chain = rhs.into();
-                                let args = &mut idx_val.into_fn_call_args();
-
                                 #[cfg(feature = "debugging")]
                                 let reset_debugger = self.run_debugger_with_reset(
                                     scope, global, lib, this_ptr, _node, level,
                                 )?;
 
+                                let crate::ast::FnCallExpr {
+                                    name, hashes, args, ..
+                                } = f.as_ref();
+                                let rhs_chain = rhs.into();
+
+                                let offset = idx_values.len() - args.len();
+                                let call_args = &mut idx_values[offset..];
+                                let pos1 = args.get(0).map_or(Position::NONE, Expr::position);
+
                                 let result = self.make_method_call(
-                                    global, caches, lib, name, *hashes, target, args, pos, level,
+                                    global, caches, lib, name, *hashes, target, call_args, pos1,
+                                    pos, level,
                                 );
+
+                                idx_values.truncate(offset);
 
                                 #[cfg(feature = "debugging")]
                                 global.debugger.reset_status(reset_debugger);
@@ -610,21 +555,46 @@ impl Engine {
         level: usize,
         new_val: Option<(Dynamic, OpAssignment)>,
     ) -> RhaiResult {
-        let (crate::ast::BinaryExpr { lhs, rhs }, chain_type, options, op_pos) = match expr {
+        let chain_type = ChainType::from(expr);
+        let (crate::ast::BinaryExpr { lhs, rhs }, options, op_pos) = match expr {
             #[cfg(not(feature = "no_index"))]
-            Expr::Index(x, options, pos) => (x.as_ref(), ChainType::Indexing, *options, *pos),
+            Expr::Index(x, options, pos) => (x.as_ref(), *options, *pos),
             #[cfg(not(feature = "no_object"))]
-            Expr::Dot(x, options, pos) => (x.as_ref(), ChainType::Dotting, *options, *pos),
+            Expr::Dot(x, options, pos) => (x.as_ref(), *options, *pos),
             expr => unreachable!("Expr::Index or Expr::Dot expected but gets {:?}", expr),
         };
 
-        let idx_values = &mut StaticVec::new_const();
+        let idx_values = &mut FnArgsVec::new_const();
 
-        self.eval_dot_index_chain_arguments(
-            scope, global, caches, lib, this_ptr, rhs, options, chain_type, idx_values, 0, level,
-        )?;
-
-        let is_assignment = new_val.is_some();
+        match rhs {
+            // Short-circuit for simple property access: {expr}.prop
+            #[cfg(not(feature = "no_object"))]
+            Expr::Property(..) if chain_type == ChainType::Dotting => (),
+            #[cfg(not(feature = "no_object"))]
+            Expr::Property(..) => unreachable!("unexpected Expr::Property for indexing"),
+            // Short-circuit for simple method call: {expr}.func()
+            #[cfg(not(feature = "no_object"))]
+            Expr::FnCall(x, ..) if chain_type == ChainType::Dotting && x.args.is_empty() => (),
+            // Short-circuit for method call with all literal arguments: {expr}.func(1, 2, 3)
+            #[cfg(not(feature = "no_object"))]
+            Expr::FnCall(x, ..)
+                if chain_type == ChainType::Dotting && x.args.iter().all(Expr::is_constant) =>
+            {
+                idx_values.extend(x.args.iter().map(|expr| expr.get_literal_value().unwrap()))
+            }
+            // Short-circuit for indexing with literal: {expr}[1]
+            #[cfg(not(feature = "no_index"))]
+            _ if chain_type == ChainType::Indexing && rhs.is_constant() => {
+                idx_values.push(rhs.get_literal_value().unwrap())
+            }
+            // All other patterns - evaluate the arguments chain
+            _ => {
+                self.eval_dot_index_chain_arguments(
+                    scope, global, caches, lib, this_ptr, rhs, options, chain_type, idx_values, 0,
+                    level,
+                )?;
+            }
+        }
 
         match lhs {
             // id.??? or id[???]
@@ -645,24 +615,25 @@ impl Engine {
                     global, caches, lib, &mut None, obj_ptr, root, expr, rhs, options, idx_values,
                     chain_type, level, new_val,
                 )
-                .map(|(v, ..)| v)
-                .map_err(|err| err.fill_position(op_pos))
             }
             // {expr}.??? = ??? or {expr}[???] = ???
-            _ if is_assignment => unreachable!("cannot assign to an expression"),
+            _ if new_val.is_some() => unreachable!("cannot assign to an expression"),
             // {expr}.??? or {expr}[???]
             expr => {
-                let value = self.eval_expr(scope, global, caches, lib, this_ptr, expr, level)?;
+                let value = self
+                    .eval_expr(scope, global, caches, lib, this_ptr, expr, level)?
+                    .flatten();
                 let obj_ptr = &mut value.into();
                 let root = ("", expr.start_position());
+
                 self.eval_dot_index_chain_helper(
                     global, caches, lib, this_ptr, obj_ptr, root, expr, rhs, options, idx_values,
                     chain_type, level, new_val,
                 )
-                .map(|(v, ..)| if is_assignment { Dynamic::UNIT } else { v })
-                .map_err(|err| err.fill_position(op_pos))
             }
         }
+        .map(|(v, ..)| v)
+        .map_err(|err| err.fill_position(op_pos))
     }
 
     /// Evaluate a chain of indexes and store the results in a [`StaticVec`].
@@ -678,7 +649,7 @@ impl Engine {
         expr: &Expr,
         parent_options: ASTFlags,
         _parent_chain_type: ChainType,
-        idx_values: &mut StaticVec<ChainArgument>,
+        idx_values: &mut FnArgsVec<Dynamic>,
         size: usize,
         level: usize,
     ) -> RhaiResultOf<()> {
@@ -690,22 +661,13 @@ impl Engine {
             Expr::MethodCall(x, ..)
                 if _parent_chain_type == ChainType::Dotting && !x.is_qualified() =>
             {
-                let crate::ast::FnCallExpr { args, .. } = x.as_ref();
-
-                let (values, pos) = args.iter().try_fold(
-                    (crate::FnArgsVec::with_capacity(args.len()), Position::NONE),
-                    |(mut values, mut pos), expr| {
-                        let (value, arg_pos) =
-                            self.get_arg_value(scope, global, caches, lib, this_ptr, expr, level)?;
-                        if values.is_empty() {
-                            pos = arg_pos;
-                        }
-                        values.push(value.flatten());
-                        Ok::<_, crate::RhaiError>((values, pos))
-                    },
-                )?;
-
-                idx_values.push(ChainArgument::from_fn_call_args(values, pos));
+                for arg_expr in x.args.as_ref() {
+                    idx_values.push(
+                        self.get_arg_value(scope, global, caches, lib, this_ptr, arg_expr, level)?
+                            .0
+                            .flatten(),
+                    );
+                }
             }
             #[cfg(not(feature = "no_object"))]
             Expr::MethodCall(..) if _parent_chain_type == ChainType::Dotting => {
@@ -713,9 +675,7 @@ impl Engine {
             }
 
             #[cfg(not(feature = "no_object"))]
-            Expr::Property(.., pos) if _parent_chain_type == ChainType::Dotting => {
-                idx_values.push(ChainArgument::Property(*pos))
-            }
+            Expr::Property(..) if _parent_chain_type == ChainType::Dotting => (),
             Expr::Property(..) => unreachable!("unexpected Expr::Property for indexing"),
 
             Expr::Index(x, options, ..) | Expr::Dot(x, options, ..)
@@ -723,34 +683,27 @@ impl Engine {
             {
                 let crate::ast::BinaryExpr { lhs, rhs, .. } = x.as_ref();
 
+                let mut _arg_values = FnArgsVec::new_const();
+
                 // Evaluate in left-to-right order
-                let lhs_arg_val = match lhs {
+                match lhs {
                     #[cfg(not(feature = "no_object"))]
-                    Expr::Property(.., pos) if _parent_chain_type == ChainType::Dotting => {
-                        ChainArgument::Property(*pos)
-                    }
+                    Expr::Property(..) if _parent_chain_type == ChainType::Dotting => (),
                     Expr::Property(..) => unreachable!("unexpected Expr::Property for indexing"),
 
                     #[cfg(not(feature = "no_object"))]
                     Expr::MethodCall(x, ..)
                         if _parent_chain_type == ChainType::Dotting && !x.is_qualified() =>
                     {
-                        let crate::ast::FnCallExpr { args, .. } = x.as_ref();
-
-                        let (values, pos) = args.iter().try_fold(
-                            (crate::FnArgsVec::with_capacity(args.len()), Position::NONE),
-                            |(mut values, mut pos), expr| {
-                                let (value, arg_pos) = self.get_arg_value(
-                                    scope, global, caches, lib, this_ptr, expr, level,
-                                )?;
-                                if values.is_empty() {
-                                    pos = arg_pos
-                                }
-                                values.push(value.flatten());
-                                Ok::<_, crate::RhaiError>((values, pos))
-                            },
-                        )?;
-                        ChainArgument::from_fn_call_args(values, pos)
+                        for arg_expr in x.args.as_ref() {
+                            _arg_values.push(
+                                self.get_arg_value(
+                                    scope, global, caches, lib, this_ptr, arg_expr, level,
+                                )?
+                                .0
+                                .flatten(),
+                            );
+                        }
                     }
                     #[cfg(not(feature = "no_object"))]
                     Expr::MethodCall(..) if _parent_chain_type == ChainType::Dotting => {
@@ -761,13 +714,14 @@ impl Engine {
                         unreachable!("invalid dot expression: {:?}", expr);
                     }
                     #[cfg(not(feature = "no_index"))]
-                    _ if _parent_chain_type == ChainType::Indexing => self
-                        .eval_expr(scope, global, caches, lib, this_ptr, lhs, level)
-                        .map(|v| {
-                            ChainArgument::from_index_value(v.flatten(), lhs.start_position())
-                        })?,
+                    _ if _parent_chain_type == ChainType::Indexing => {
+                        _arg_values.push(
+                            self.eval_expr(scope, global, caches, lib, this_ptr, lhs, level)?
+                                .flatten(),
+                        );
+                    }
                     expr => unreachable!("unknown chained expression: {:?}", expr),
-                };
+                }
 
                 // Push in reverse order
                 let chain_type = expr.into();
@@ -777,7 +731,9 @@ impl Engine {
                     size, level,
                 )?;
 
-                idx_values.push(lhs_arg_val);
+                if !_arg_values.is_empty() {
+                    idx_values.extend(_arg_values);
+                }
             }
 
             #[cfg(not(feature = "no_object"))]
@@ -786,8 +742,8 @@ impl Engine {
             }
             #[cfg(not(feature = "no_index"))]
             _ if _parent_chain_type == ChainType::Indexing => idx_values.push(
-                self.eval_expr(scope, global, caches, lib, this_ptr, expr, level)
-                    .map(|v| ChainArgument::from_index_value(v.flatten(), expr.start_position()))?,
+                self.eval_expr(scope, global, caches, lib, this_ptr, expr, level)?
+                    .flatten(),
             ),
             _ => unreachable!("unknown chained expression: {:?}", expr),
         }
@@ -796,7 +752,6 @@ impl Engine {
     }
 
     /// Call a get indexer.
-    /// [`Position`] in [`EvalAltResult`] may be [`NONE`][Position::NONE] and should be set afterwards.
     #[inline(always)]
     fn call_indexer_get(
         &self,
@@ -820,7 +775,6 @@ impl Engine {
     }
 
     /// Call a set indexer.
-    /// [`Position`] in [`EvalAltResult`] may be [`NONE`][Position::NONE] and should be set afterwards.
     #[inline(always)]
     fn call_indexer_set(
         &self,
@@ -1041,6 +995,11 @@ impl Engine {
                     value: ch.into(),
                     index: offset,
                 })
+            }
+
+            #[cfg(not(feature = "no_closure"))]
+            Dynamic(Union::Shared(..)) => {
+                unreachable!("`get_indexed_mut` cannot handle shared values")
             }
 
             _ if use_indexers => self